Fade working again.  Overlaid tracked views working with async rendering.  Drawable atlases now respecting draw priority and visibility range.  Fixed a but in single byte compressed textures in texture atlases.  A few more changes related to the great double migration.  Now rescaling texture to include a boundary, thus fixing the boundary problem in texture atlases.  Dealing with view resize failing in EAGLView.  Loft layer now has a special mode for filled, layered polys.  Fixed an ordering problem with skirts.  Labels working again.
Former-commit-id: eb4d969e2d305f61463a015ec6bc3c75b1496a5c

+++ WhirlyGlobeSrc/WhirlyGlobeLib/include/WhirlyKitView.h
@@ -93,7 +93,7 @@ typedef std::set<NSObject<WhirlyKitViewWatcherDelegate> * __weak> WhirlyKitViewW
 - (Eigen::Matrix4d)calcProjectionMatrix:(WhirlyKit::Point2f)frameBufferSize margin:(float)margin;
 
 /// Return the nominal height above the surface of the data
-- (float)heightAboveSurface;
+- (double)heightAboveSurface;
 
 /// From a screen point calculate the corresponding point in 3-space
 - (WhirlyKit::Point3d)pointUnproject:(WhirlyKit::Point2f)screenPt width:(unsigned int)frameWidth height:(unsigned int)frameHeight clip:(bool)clip;
